Now compiling on Mac OSX, clients work, vision doesn't.
In this commit some modifications have been made so the code compiles on
Mac OSX, one still needs to compile locally dependencies such as
protobuf.
The clients (client, graphicalClient and logClient) seem to work
correctly, libdc1394 is disabled if not found which should simplify the
build process when its not available, although compiling without
libdc1394 vision segfaults, there is still some debugging to do in order
to fix this bug.