Moving robocup-ssl-teams.xml to /bin dir.
In order not to pollute the root of the project its been moved to the
bin folder, the standard manner should be going to the bin folder and
running from there.
Also, .gitignore will ignore changes to that file so one can customize
it and it doesn't show up on modified files, because it's a local
setting, when changes to that file should be commited use the -f flag.